// client-monorepo/libs/shared/common/service-data/src/lib/data-access/models/app-service-category-names.enum.ts
export enum AppServiceCategoryNamesEnum {
  PAY = 'payment',
  MOBILE = 'mobile',
  WEALTH = 'wealth',
  INSURANCE = 'insurance',
  VEHICLE = 'vehicle',
  INSTALLMENT = 'installment',
  SUGGESTED_SERVICES = 'suggested_services',
  OTHER_SERVICES = 'other_services',
  BNPL_SERVICES = 'bnpl_services',
  BILL = 'bill',
  BNPL_CREDIT_SERVICES = 'bnpl_credit_services',
  WALLET = 'wallet',
}